EZP-31853: Fixed inline tags crashing script on creation / edit
Inline custom tags expect to work with selection. However, in cases where there were no selection or initially
that custom tag was not an inline tag, they would crash the script and cause the editor to become unresponsive.
By adding a check if any node was found this is circumvented.