[PHPStan] Added new issues to PHPStan baseline after the release (#234)
PHPStan v1.10.22 uncovered new pre-existing issues related to DOM extension.

The issue is that `\DOMXPath::query` declares that it returns `DOMNodeList` which contains `DOMNode` instances, but it actually returns a list which contains `DOMElement` instances. It requires further investigation if it's a bug in DOM or if there's a case when it returns actually `DOMNode` and shouldn't be passed to the method.